Rename report components to align with entryTypes
