+++ Apps/WebClient/src/NewClientApp/src/components/private/reports/LabResultReportComponent.vue
