+++ Apps/WebClient/src/NewClientApp/src/components/private/reports/MedicationReportComponent.vue
