+++ Apps/WebClient/src/NewClientApp/src/components/private/reports/NoteReportComponent.vue
